docs: add description to functions
