+++ cerror/cerror.go
@@ -4,36 +4,27 @@ import (
 	"github.com/pkg/errors"
 )
 
-type ErrorType uint
-
-const (
-	NoType ErrorType = iota
-	ValidationError
-	DuplicatedError
-	NotFoundError
-)
-
-type customError struct {
-	errorType     ErrorType
-	originalError error
-}
-
+// Create a validation error
 func NewValidationError(err error) error {
 	return customError{errorType: ValidationError, originalError: err}
 }
 
+// Create a Duplicated error
 func NewDuplicatedError(err error) error {
 	return customError{errorType: DuplicatedError, originalError: err}
 }
 
+// Create a Not Found error
 func NewNotFoundError(err error) error {
 	return customError{errorType: NotFoundError, originalError: err}
 }
 
-func (error customError) Error() string {
-	return error.originalError.Error()
+// Get error message from a custom error
+func (err customError) Error() string {
+	return err.originalError.Error()
 }
 
+// Check if given error has the "Validation Error" type
 func IsValidationError(err error) bool {
 	if customErr, ok := errors.Cause(err).(customError); ok {
 		return customErr.errorType == ValidationError
@@ -42,6 +33,7 @@ func IsValidationError(err error) bool {
 	return false
 }
 
+// Check if given error has the "Is Duplicated Error" type
 func IsDuplicatedError(err error) bool {
 	if customErr, ok := errors.Cause(err).(customError); ok {
 		return customErr.errorType == DuplicatedError
@@ -50,6 +42,7 @@ func IsDuplicatedError(err error) bool {
 	return false
 }
 
+// Check if given error has the "Not Found Error" type
 func IsNotFoundError(err error) bool {
 	if customErr, ok := errors.Cause(err).(customError); ok {
 		return customErr.errorType == NotFoundError
